requirements.txt
alembic==1.16.4
APScheduler==3.11.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
Flask==3.1.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==4.0.1
marshmallow-sqlalchemy==1.4.2
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.1.1
requests==2.32.5
SQLAlchemy==2.0.42
typing_extensions==4.14.1
tzlocal==5.3.1
urllib3==2.5.0
Werkzeug==3.1.3
